inst/rpyants/requirements.txt
antspyx>=0.3.8
nibabel>=5.0.1
